🔨Refactor: Social Entity 수정
CustomOAuth2ClientProperties 위치 수정
CustomOAuth2Provider 위치 수정
JwtDto 위치 수정
JwtProvider 기능을 JwtPlugin에 병합
SocialJwtAuthenticationFilter 삭제
